fix: search global node_modules with interactive shell
